GDB-11605 Fix cluster-joining on scale up
Scaling up does not connect the new instance to the cluster.

When scaling up the GraphDB instances, for example, from 3 to 4 (node_count=3 → 4), the fourth machine will not join the cluster.

The solution is to update the environment variable that stores the admin password, as it differs from the one used in the other scripts.